requirements.txt
pytest==7.4.0
requests==2.31.0
PyYAML==6.0.1
allure-pytest==2.13.2
python-dotenv==1.0.0
pytest-html==3.2.0
pytest-metadata==3.0.0
